requirements.txt
langchain==1.0.2
langchain-core==1.0.1
langchain-community==0.3.0
langchain-google-genai==2.0.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
streamlit==1.40.0
pydantic==2.12.3
google-generativeai==0.8.3